ml_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic[email]==2.9.0
pydantic-settings==2.5.2
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
python-dotenv==1.0.1
